MDD-22 update : update and delete

+++ todo/views.py
@@ -1,20 +1,17 @@
-# from django.shortcuts import render ?????
-
 from .models import Todo
-from .serializers import TodoSerializer
-from rest_framework.decorators import api_view
-from rest_framework.response import Response
 from rest_framework import status
+from .serializers import TodoSerializer
+from rest_framework.response import Response
+from rest_framework.decorators import api_view
 
 
 # To-do list 할일 표시
 @api_view(['GET'])
 def todo_list(request):
-    select_date = request.query_params.get('select_date')  # 쿼리 파라미터로 날짜 필터링
-    if select_date:
-        todo = Todo.objects.filter(user=request.user, select_date=select_date).order_by('-created_at')
-    else:
-        todo = Todo.objects.filter(user=request.user).order_by('-created_at')
+
+    todo = Todo.objects.filter(user=request.user, select_date=select_date).order_by('-created_at')
+
+    todo = Todo.objects.filter(user=request.user).order_by('-created_at')
 
     serializer = TodoSerializer(todo, many=True)
     return Response(serializer.data, status=status.HTTP_200_OK)
@@ -31,31 +28,28 @@ def todo_create(request):
 
 
 
-# 작성 제거
-@api_view(['DELETE'])
-def todo_delete(request, todo_id):
-    try:
-        todo = Todo.objects.get(pk=todo_id, user=request.user) # 유저의 할 일만 삭제 가능하도록
-    except Todo.DoesNotExist:
-        return Response({"error": "Todo not found"}, status=status.HTTP_404_NOT_FOUND)
-    
-    todo.delete()
-    return Response({"message": "Todo list deleted successfully."}, status=status.HTTP_204_NO_CONETENT)
-
-
 # 작성 수정
-@api_view(['PUT'])
+@api_view(['PUT','DELETE'])
 def todo_update(request, todo_id):
-    try:
-        todo = Todo.objects.get(pk=todo_id, user = request.user) # 유저의 할 일만 수정 가능하도록
-    except Todo.DoesNotExist:
-        return Response({"error": "Todo not found"}, status=status.HTTP_404_NOT_FOUND)
+    if request.method == 'PUT':
+        try:
+            todo = Todo.objects.get(pk=todo_id, user = request.user) # 유저의 할 일만 수정 가능하도록
+        except Todo.DoesNotExist:
+            return Response({"error": "Todo not found"}, status=status.HTTP_404_NOT_FOUND)
 
-    serializer = TodoSerializer(todo, data=request.data)
-    if serializer.is_valid():
-        serializer.save()
-        return Response(serializer.data, status=status.HTTP_200_OK)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        serializer = TodoSerializer(todo, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    
 
-
+    if request.method == 'DELETE':
+        try:
+            todo = Todo.objects.get(pk=todo_id, user=request.user) # 유저의 할 일만 삭제 가능하도록
+        except Todo.DoesNotExist:
+            return Response({"error": "Todo not found"}, status=status.HTTP_404_NOT_FOUND)
+        
+        todo.delete()
+        return Response({"message": "Todo list deleted successfully."}, status=status.HTTP_200_OK)
